Move specialize module to type_checker and refactor parser
The specialize module has been moved from core/src/parser to core/src/type_checker, and related imports and module declarations have been updated accordingly. Intrinsic operator definitions and constants were relocated from op.rs to parser.rs. Type handling logic was removed from the parser, and function signatures were updated to reflect these changes.
