+++ core/src/parser/mod.rs
@@ -3,5 +3,4 @@ pub mod error;
 pub mod op;
 pub mod parser;
 pub mod pattern;
-pub mod specialize;
 mod tests;
