+++ core/src/type_checker/mod.rs
@@ -1,4 +1,5 @@
 pub mod checked_ast;
 pub mod checker;
+mod specialize;
 mod tests;
 pub mod types;
